Updated data field to floating point, renamed app

+++ PowerHorses/greenhouse_monitoring/models.py
@@ -3,9 +3,9 @@ from django.db import models
 # Create your models here.
 
 class parameters(models.Model):
-	temperature = models.IntegerField()
-	humidity = models.IntegerField()
-	light_intensity = models.IntegerField()
+	temperature = models.FloatField()
+	humidity = models.FloatField()
+	light_intensity = models.FloatField()
 	timestamp = models.DateTimeField(auto_now=True)
 
 	class Meta:
